+++ src/redis/types/command_token.rs
@@ -18,5 +18,5 @@ pub enum CommandToken {
     Discard,
     Info(String),
     ReplConf(Vec<String>),
-    Psync(String, i32),
+    Psync(String, i64),
 }
